Complete rewrite for version 2 (#4)
This is a complete rewrite. Instead of checking or running gleam code via checkmark, checkmark can now either check that a snippet matches the content in a file, or update the markdown file with the content.

Dependencies have also been simplified, and the parsing of the markdown file no longer uses the package it used to, as it had some issues in it, and doesn't seem to be actively maintained.